+++ typeddfs/typed_dfs.py
@@ -158,17 +158,17 @@ class TypedDf(BaseDf):
             return self.__class__.convert(df)
 
     @classmethod
-    def read_csv(cls, path: PathLike, *args, **kwargs) -> __qualname__:
-        df = pd.read_csv(Path(path), index_col=False, **kwargs)
+    def read_csv(cls, filepath_or_buffer: PathLike, *args, **kwargs) -> __qualname__:
+        df = pd.read_csv(filepath_or_buffer, index_col=False, **kwargs)
         return cls.convert(df)
 
-    def to_csv(self, path: PathLike, *args, **kwargs) -> Optional[str]:
+    def to_csv(self, path_or_buf: PathLike, *args, **kwargs) -> Optional[str]:
         # TODO not checking for index in the args
         if "index" in kwargs:
-            return super().to_csv(path, *args, **kwargs)
+            return super().to_csv(path_or_buf, *args, **kwargs)
         else:
             df = self.vanilla().reset_index()
-            return df.to_csv(path, *args, index=False, **kwargs)
+            return df.to_csv(path_or_buf, *args, index=False, **kwargs)
 
     @classmethod
     def is_valid(cls, df: pd.DataFrame) -> bool: